Fix architecture : separate components & pages
