+++ src/app/modules/app-routing/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 
-import { HomeComponent } from '../../components/home/home.component';
+import { HomeComponent } from '../../pages/home/home.component';
 
 const appRoutes: Routes = [
   { path: '', component: HomeComponent },
